Revert "feat: add registration form (#48)"
This reverts commit 7560ecbba17f3b77a6cb465e4b046b3b65666cc5.
