+++ flourish-client/src/App.tsx
@@ -1,4 +1,3 @@
-import { useEffect } from "react";
 import { Box } from "@chakra-ui/react";
 import { AnimatePresence } from "framer-motion";
 import { Route, Routes, useLocation } from "react-router-dom";
@@ -11,16 +10,9 @@ import Auth from "./pages/auth";
 import SignIn from "./pages/auth/signin";
 import SignUp from "./pages/auth/signup";
 import AdminDashboard from "./pages/dashboard/admin";
-import OverView from "./pages/dashboard/admin/overview";
 import Questionnaire from "./pages/dashboard/admin/questionnaire";
 import QuestionnaireDetails from "./pages/dashboard/admin/questionnaire/QuestionnaireDetails";
 import QuestionnaireList from "./pages/dashboard/admin/questionnaire/QuestionnaireList";
-import ClientDashboard from "./pages/dashboard/client";
-import Forms from "./pages/dashboard/client/form";
-import RegistrationFormDetails from "./pages/dashboard/client/form/RegistrationFormDetails";
-import RegistrationFormFillUp from "./pages/dashboard/client/form/RegistrationFormFillUp";
-import RegistrationFormList from "./pages/dashboard/client/form/RegistrationFormList";
-import CounselorDashboard from "./pages/dashboard/counselor";
 import Homepage from "./pages/homepage";
 import ClientDashboard from "./pages/dashboard/client";
 import CounselorDashboard from "./pages/dashboard/counselor";
@@ -52,17 +44,13 @@ const App = () => {
           <Route index element={<h1>Members</h1>} />
           <Route path={nav.members} element={<h1>Members</h1>} />
           <Route path={nav.sessionRequest} element={<h1>sessionRequest</h1>} />
-          <Route path={nav.overview} element={<OverView />} />
+          <Route path={nav.overview} element={<h1>Overview</h1>} />
         </Route>
       );
     else if (client)
       return (
         <Route path={`${nav.dashboard}`} element={<ClientDashboard />}>
-          <Route path={nav.form} element={<Forms />}>
-            <Route index element={<RegistrationFormList />} />
-            <Route path="create" element={<RegistrationFormFillUp />} />
-            <Route path=":id" element={<RegistrationFormDetails />} />
-          </Route>
+          <Route path={nav.form} element={<Form />} />
         </Route>
       );
     else if (counselor)
@@ -96,7 +84,6 @@ const App = () => {
           )}
 
           {isSignedIn && dashboardRoutes()}
-          <Route path={nav[404]} element={<NotFound />} />
           <Route path="*" element={<NotFound />} />
         </Routes>
       </AnimatePresence>
