+++ flourish-client/src/components/questionnaire/QuestionnaireCard.tsx
@@ -1,36 +1,24 @@
 import { useState } from "react";
 import {
-  Box,
   Button,
   Card,
   CardBody,
   CardFooter,
   CardHeader,
-  Flex,
   Heading,
   Text,
   VStack,
-  chakra,
-  shouldForwardProp,
 } from "@chakra-ui/react";
-import { AnimatePresence, isValidMotionProp, motion } from "framer-motion";
-import { IoCreateOutline, IoTrashOutline } from "react-icons/io5";
-import { VscChevronDown } from "react-icons/vsc";
 import { Link } from "react-router-dom";
-
 import { routes } from "../../assets/data/routes";
+
 import { useAppDispatch } from "../../hooks/useStore";
 import {
   fetchQuestionnaire,
   removeQuestionnaire,
 } from "../../store/actions/questionnaireActions";
-import theme from "../../theme/theme";
 import { Questionnaire, QuestionnaireKeys } from "../../types/Questionnaire";
-
-const ChakraBox = chakra(motion.div, {
-  shouldForwardProp: (prop) =>
-    isValidMotionProp(prop) || shouldForwardProp(prop),
-});
+import ButtonFull from "../common/button/ButtonFull";
 
 const QuestionnaireCard = ({
   questionnaire,
@@ -39,13 +27,9 @@ const QuestionnaireCard = ({
 }) => {
   const dispatch = useAppDispatch();
   const [isLoading, setIsLoading] = useState(false);
-  const [isExpanded, setIsExpanded] = useState(false);
-  const baseQuestions = 3;
 
   const renderedQuestionnaire = () => {
-    const maxQuestions = isExpanded
-      ? questionnaire.questionnaireFields.length
-      : baseQuestions;
+    const maxQuestions = 3;
 
     const questions = questionnaire.questionnaireFields.map(
       ({ question }, index) => {
@@ -53,7 +37,6 @@ const QuestionnaireCard = ({
         return (
           <Text
             key={index}
-            as={motion.p}
             textTransform={"capitalize"}
             fontSize={"lg"}
             color={"font.muted"}
@@ -61,10 +44,14 @@ const QuestionnaireCard = ({
             style={{
               lineClamp: 1,
             }}
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
           >
-            {question}
+            {index !== maxQuestions - 1 ? (
+              question
+            ) : (
+              <Text as={"span"} letterSpacing={"widest"}>
+                ...
+              </Text>
+            )}
           </Text>
         );
       }
@@ -78,120 +65,61 @@ const QuestionnaireCard = ({
   };
 
   return (
-    <AnimatePresence initial={false}>
-      <ChakraBox
-        layout
-        //@ts-ignore
-        transition={{
-          layout: {
-            type: "spring",
-            bounce: 0.3,
-            opacity: { delay: 0.2 },
-          },
-        }}
-      >
-        <Card
-          px={24}
-          py={16}
-          border={"2px solid #e9f2fd"}
-          borderRadius={"xl"}
-          boxShadow={"none"}
+    <Card
+      px={16}
+      py={8}
+      border={"2px solid #e9f2fd"}
+      borderRadius={"xl"}
+      boxShadow={"none"}
+    >
+      <CardHeader pb={2}>
+        <Heading
+          size="2xl"
+          color={"font.general"}
+          fontWeight={"semibold"}
+          textTransform={"capitalize"}
+          letterSpacing={"tight"}
         >
-          <CardHeader
-            as={motion.div}
-            layout={"position"}
-            display={"flex"}
-            pb={2}
-            justifyContent={"space-between"}
-          >
-            <Heading
-              size="2xl"
-              color={"font.general"}
-              fontWeight={"semibold"}
-              textTransform={"capitalize"}
-              letterSpacing={"tight"}
-            >
-              {questionnaire.name}
-            </Heading>
-            <Flex gap={24}>
-              <Button
-                as={Link}
-                to={`${routes.questionnaire}/${questionnaire.id}`}
-                boxSize={"36px"}
-                borderRadius={"lg"}
-                variant={"outline"}
-                colorScheme={"linkedin"}
-              >
-                <IoCreateOutline fontSize={"20px"} />
-              </Button>
-              <Button
-                boxSize={"36px"}
-                borderRadius={"lg"}
-                variant={"outline"}
-                colorScheme={"red"}
-                isLoading={isLoading}
-                onClick={async () => {
-                  setIsLoading(true);
-                  const responseRemove = await dispatch(
-                    removeQuestionnaire({
-                      id: questionnaire[QuestionnaireKeys.ID] as string,
-                    })
-                  );
-                  const responseFetch = await dispatch(fetchQuestionnaire());
-                  if (responseRemove.payload && responseFetch.payload) {
-                    setIsLoading(false);
-                  }
-                }}
-              >
-                <IoTrashOutline fontSize={"20px"} />
-              </Button>
-            </Flex>
-          </CardHeader>
-
-          <CardBody>{renderedQuestionnaire()}</CardBody>
-          {questionnaire.questionnaireFields.length > baseQuestions && (
-            <CardFooter justifyContent={"center"} p={0}>
-              <Box
-                as={motion.div}
-                fontSize={28}
-                initial="visible"
-                animate={"visible"}
-                whileHover="hover"
-                variants={downArrowVariants}
-                custom={isExpanded}
-                cursor={"pointer"}
-                onClick={() => setIsExpanded((prev) => !prev)}
-                layout={"preserve-aspect"}
-              >
-                <VscChevronDown />
-              </Box>
-            </CardFooter>
-          )}
-        </Card>
-      </ChakraBox>
-    </AnimatePresence>
+          {questionnaire.name}
+        </Heading>
+      </CardHeader>
+      <CardBody>{renderedQuestionnaire()}</CardBody>
+      <CardFooter gap={24}>
+        <ButtonFull
+          as={Link}
+          to={`${routes.questionnaire}/${questionnaire.id}`}
+          px={20}
+          py={18}
+          fontSize={"lg"}
+        >
+          View Details
+        </ButtonFull>
+        <Button
+          px={20}
+          py={17}
+          fontSize={"lg"}
+          variant={"outline"}
+          borderRadius={"xl"}
+          colorScheme={"red"}
+          isLoading={isLoading}
+          onClick={async () => {
+            setIsLoading(true);
+            const responseRemove = await dispatch(
+              removeQuestionnaire({
+                id: questionnaire[QuestionnaireKeys.ID] as string,
+              })
+            );
+            const responseFetch = await dispatch(fetchQuestionnaire());
+            if (responseRemove.payload && responseFetch.payload) {
+              setIsLoading(false);
+            }
+          }}
+        >
+          Delete
+        </Button>
+      </CardFooter>
+    </Card>
   );
 };
 
 export default QuestionnaireCard;
-
-const transitionTween = {
-  duration: 0.3,
-  type: "tween",
-};
-
-const downArrowVariants = {
-  visible: (isExpanded: boolean) => ({
-    rotate: isExpanded ? 180 : 360,
-    scale: 1,
-    color: theme.colors.font.muted2,
-    transition: { ...transitionTween },
-  }),
-  hover: {
-    scale: 1.1,
-    color: theme.colors.font.primary,
-    transition: {
-      ...transitionTween,
-    },
-  },
-};
