+++ flourish-client/src/pages/dashboard/admin/questionnaire/index.tsx
@@ -1,12 +1,15 @@
+import { Box, SimpleGrid } from "@chakra-ui/react";
 import { useEffect } from "react";
-import { Box } from "@chakra-ui/react";
-import { Outlet } from "react-router-dom";
-
-import { useAppDispatch } from "../../../../hooks/useStore";
+import { useAppDispatch, useAppSelector } from "../../../../hooks/useStore";
 import { fetchQuestionnaire } from "../../../../store/actions/questionnaireActions";
+import QuestionnaireCard from "../../../../components/questionnaire/QuestionnaireCard";
+import { Outlet } from "react-router-dom";
 
 const Questionnaire = () => {
   const dispatch = useAppDispatch();
+  const questionnaires = useAppSelector(
+    (state) => state.questionnaire.questionnaires
+  );
 
   useEffect(() => {
     dispatch(fetchQuestionnaire());
