--- flourish-client/src/pages/dashboard/client/form/RegistrationFormFillUp.tsx
@@ -1,7 +1,0 @@
-import RegistrationForm from "../../../../components/form/RegistrationForm";
-
-const RegistrationFormFillUp = () => {
-  return <RegistrationForm />;
-};
-
-export default RegistrationFormFillUp;
