--- flourish-client/src/utils/date.ts
@@ -1,7 +1,0 @@
-export const stringToDate = (date: string) => {
-  return new Date(date).toLocaleDateString("en-US", {
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-  });
-};
